NIT Refactor `Fmt` & `FmtTests`   - fill in missing type annotations   - extract `FmtOptionsOps` to a separate util source file
